refactor(ui): extract GlossarySelect from AddTermsModal into sharedV2
